make false positive warnings go away
issue described in this issue https://github.com/expo/router/issues/834
