+++ app/auth/login.tsx
@@ -1,5 +1,4 @@
 import { Text, Button, TextInput, Center } from "@/components";
-import { Stack } from "expo-router";
 import { View, TouchableWithoutFeedback, Keyboard } from "react-native";
 import { useState } from "react";
 import { useAuth } from "@/hooks";
@@ -28,38 +27,35 @@ export default function Login() {
   };
 
   return (
-    <>
-      <Stack.Screen options={{ headerTitle: "Login" }} />
-      <TouchableWithoutFeedback onPress={() => Keyboard.dismiss()}>
-        <Center style={{ paddingHorizontal: 24 }}>
-          <View style={{ marginBottom: 24, width: "100%" }}>
-            <TextInput
-              label="nostr nsec"
-              secureTextEntry
-              autoCorrect={false}
-              value={nsec}
-              onChangeText={(value) => {
-                setNsec(value);
-                setErrorMessage("");
-              }}
-              errorMessage={errorMessage}
-            />
-          </View>
-          <Button
-            onPress={handleLogin}
-            disabled={isLoggingIn}
-            loading={isLoggingIn}
-          >
-            Register
-          </Button>
-          <View style={{ marginTop: 80 }}>
-            <Text style={{ fontSize: 18 }}>
-              Your private key will only be stored on your device and not on
-              Wavlake systems. Wavlake will never have access to your key.
-            </Text>
-          </View>
-        </Center>
-      </TouchableWithoutFeedback>
-    </>
+    <TouchableWithoutFeedback onPress={() => Keyboard.dismiss()}>
+      <Center style={{ paddingHorizontal: 24 }}>
+        <View style={{ marginBottom: 24, width: "100%" }}>
+          <TextInput
+            label="nostr nsec"
+            secureTextEntry
+            autoCorrect={false}
+            value={nsec}
+            onChangeText={(value) => {
+              setNsec(value);
+              setErrorMessage("");
+            }}
+            errorMessage={errorMessage}
+          />
+        </View>
+        <Button
+          onPress={handleLogin}
+          disabled={isLoggingIn}
+          loading={isLoggingIn}
+        >
+          Register
+        </Button>
+        <View style={{ marginTop: 80 }}>
+          <Text style={{ fontSize: 18 }}>
+            Your private key will only be stored on your device and not on
+            Wavlake systems. Wavlake will never have access to your key.
+          </Text>
+        </View>
+      </Center>
+    </TouchableWithoutFeedback>
   );
 }
